rename memory-redis to memory-upstash-redis
