+++ packages/memory-redis/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@aigur/memory-redis",
+	"name": "@aigur/memory-upstash-redis",
 	"version": "0.1.0",
 	"description": "",
 	"main": "dist/index.js",
